requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
numpy
pandas
scikit-learn
matplotlib
seaborn
pyyaml
opencv-python-headless
tensorboard
